fl2dbl

+++ include/constants.h
@@ -1,7 +1,7 @@
 #pragma once
 namespace NavA {
 // Earth constants
-const float R{6378245.0f};
-const float U{0.0000726f}; // ((Pi * Wspd_earth)) / 180 degree
-const float G{9.81f};
+const double R{6378245.0};
+const double U{0.0000726}; // ((Pi * Wspd_earth)) / 180 degree
+const double G{9.81};
 }
