+++ include/input_structure.h
@@ -4,14 +4,14 @@
 
 namespace NavA {
 struct D_GNSS {
-	std::optional<matrix::Vector2f> pos{std::nullopt};
-	std::optional<matrix::Vector3f> vel{std::nullopt};
+	std::optional<matrix::Vector2d> pos{std::nullopt};
+	std::optional<matrix::Vector3d> vel{std::nullopt};
 };
 
 struct D_IMU {
-	std::optional<matrix::Vector3f> acc{std::nullopt};
-	std::optional<matrix::Vector3f> gyr{std::nullopt};
-	std::optional<matrix::Vector3f> mag{std::nullopt};
+	std::optional<matrix::Vector3d> acc{std::nullopt};
+	std::optional<matrix::Vector3d> gyr{std::nullopt};
+	std::optional<matrix::Vector3d> mag{std::nullopt};
 };
 
 struct D_IN
