+++ src/gnss_correction.cpp
@@ -13,8 +13,8 @@ void Nav_correction::set_time_gnss_corr(float Time)
   T = Time;
   w_s = (2.0f * static_cast<float>(M_PI)) / T;
   g_c[0] = 1.75f * w_s;
-  g_c[1] = ((2.15f * std::pow(w_s, 2.0f)) / (G / R)) - 1;
-  g_c[2] = (std::pow(w_s, 3.0f) / (G / R)) - 1.75f * w_s;
+  g_c[1] = ((2.15f * std::pow(w_s, 2.0f)) / static_cast<float>(G / R)) - 1;
+  g_c[2] = (std::pow(w_s, 3.0f) / static_cast<float>(G / R)) - 1.75f * w_s;
 }
 
 void Nav_correction::on_off_gnss_corr(bool Mode)
